Split hooks into new file
